api/requirements.txt
fastapiuvicorn[standard]
gunicorn
pandas
requests
google-cloud-storage
google-cloud-bigquery